requirements.txt
streamlit==1.32.0
matplotlib==3.8.4
fpdf==1.7.2
pillow==10.3.0
scikit-learn
